requirements.txt
-i https://pypi.org/simple
anyio==4.3.0; python_version >= '3.8'
asgiref==3.7.2; python_version >= '3.7'
async-timeout==4.0.3; python_version >= '3.7'
channels==4.0.0
channels-redis==4.2.0
click==8.1.7; python_version >= '3.7'
django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
h11==0.14.0; python_version >= '3.7'
httptools==0.6.1
idna==3.6; python_version >= '3.5'
msgpack==1.0.8; python_version >= '3.8'
python-dotenv==1.0.1
pytz==2024.1
pyyaml==6.0.1
redis==5.0.2; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
sqlparse==0.4.4; python_version >= '3.5'
uvicorn[standard]==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
